+++ ion/nn/__init__.py
@@ -2,7 +2,7 @@
 
 from .blocks.mlp import MLP
 from .blocks.sequential import Sequential
-from .buffer import BufferModule, Buffers
+from .buffer import Buffer
 from .layers.attention import CrossAttention, SelfAttention
 from .layers.conv import Conv, ConvTranspose
 from .layers.dropout import Dropout
@@ -20,8 +20,7 @@ from .param import Param
 
 __all__ = [
     "Module",
-    "BufferModule",
-    "Buffers",
+    "Buffer",
     "Param",
     "AvgPool",
     "BatchNorm",